# server/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "MTC API Documentation",
    "version": "1.0.0",
    "title": "MTC API Documentation"
  },
  "host": "localhost:10010",
  "basePath": "/api",
  "schemes": [
    "http",
    "https"
  ],
  "parameters": {
    "clientId": {
      "name": "clientId",
      "type": "string",
      "in": "header",
      "required": true,
      "description": "APIs are secured by clientId."
    },
    "clientSecret": {
      "name": "clientSecret",
      "type": "string",
      "in": "header",
      "required": true,
      "description": "clientSecret"
    },
    "guest": {
      "name": "guest",
      "type": "string",
      "in": "header",
      "required": true,
      "description": "guest id"
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/user/details": {
      "parameters": [
        {
          "$ref": "#/parameters/clientId"
        },
        {
          "$ref": "#/parameters/clientSecret"
        }
      ],
      "post": {
        "tags": [
          "User"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "API for add user Details",
        "consumes": [
          "application/json"
        ],
        "operationId": "addUserDetails",
        "description": "API for add Details\n",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "phone",
                "name",
                "pinCode",
                "mapLocation"
              ],
              "properties": {
                "name": {
                  "type": "string"
                },
                "phone": {
                  "type": "string"
                },
                "pinCode": {
                  "type": "string"
                },
                "mapLocation": {
                  "type": "string"
                },
                "shopName": {
                  "type": "string"
                },
                "area": {
                  "type": "string"
                },
                "landmark": {
                  "type": "string"
                },
                "state": {
                  "type": "string"
                },
                "city": {
                  "type": "string"
                },
                "country": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated Success Fully  ",
            "schema": {
              "type": "object"
            }
          }
        }
      },
      "get": {
        "tags": [
          "User"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "API for get Details",
        "consumes": [
          "application/json"
        ],
        "operationId": "getUserDetails",
        "description": "API for get Details\n",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": " get Category successfully ",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
...
    "/open/cart": {
      "parameters": [
        {
          "$ref": "#/parameters/clientId"
        },
        {
          "$ref": "#/parameters/clientSecret"
        },
        {
          "$ref": "#/parameters/guest"
        }
      ],
      "post": {
        "tags": [
          "Cart"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "API for add Cart",
        "consumes": [
          "application/json"
        ],
        "operationId": "addCart",
        "description": "API for add Cart\n",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "quantity",
                "item",
                "propertyId"
              ],
              "properties": {
                "quantity": {
                  "type": "number",
                  "minimum": 0
                },
                "item": {
                  "type": "string"
                },
                "propertyId": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": " Register user successfully ",
            "schema": {
              "type": "object"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Cart"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "API for getCart",
        "consumes": [
          "application/json"
        ],
        "operationId": "getCart",
        "description": "API for get Cart\n",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "minimum": 0,
            "required": false,
            "type": "number"
          },
          {
            "in": "query",
            "name": "id",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "item",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "limit",
            "minimum": 0,
            "maximum": 20,
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": " get Cart successfully ",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
...
    "/open/skip": {
      "parameters": [
        {
          "$ref": "#/parameters/clientId"
        },
        {
          "$ref": "#/parameters/clientSecret"
        },
        {
          "$ref": "#/parameters/guest"
        }
      ],
      "get": {
        "tags": [
          "Cart"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "API for transferCartData",
        "consumes": [
          "application/json"
        ],
        "operationId": "getCart",
        "description": "API for get Cart\n",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": " get Cart successfully ",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
...
  },
...
}